Conflicts:
	bc-netdisk/src/main/java/cn/bc/netdisk/web/struts2/NetdiskFilesAction.java

+++ bc-netdisk/src/main/java/cn/bc/netdisk/service/NetdiskFileServiceImpl.java
@@ -47,15 +47,18 @@ public class NetdiskFileServiceImpl extends DefaultCrudService<NetdiskFile>
 		// 若是文件夹且文件夹下的子文件都删除
 		if (isRelevanceDelete) {
 			// 通过id找相关的所有pid
-			Serializable[] ids = this.netdiskFileDao.getChildIdsById(id);
+			Serializable[] ids = this.netdiskFileDao
+					.getMyselfAndChildFileId(id);
 			// 如果pid中是文件类型的就删除物理文件
 			for (int i = 0; i < ids.length; i++) {
 				NetdiskFile df = this.netdiskFileDao.load(ids[i]);
-				if (df.getType() == NetdiskFile.TYPE_FILE) {
-					String path = Attach.DATA_REAL_PATH + "/netdisk/"
-							+ df.getPath();
-					File f = new File(path);
-					f.delete();
+				if (df != null) {
+					if (df.getType() == NetdiskFile.TYPE_FILE) {
+						String path = Attach.DATA_REAL_PATH + "/netdisk/"
+								+ df.getPath();
+						File f = new File(path);
+						f.delete();
+					}
 				}
 				// 批量删除文件数据
 				if (i == ids.length - 1) {
@@ -92,5 +95,13 @@ public class NetdiskFileServiceImpl extends DefaultCrudService<NetdiskFile>
 
 	public NetdiskShare getNetdiskShare(Long aid, Long pid) {
 		return this.netdiskFileDao.getNetdiskShare(aid, pid);
+	}
+
+	public Serializable[] getMyselfAndChildFileId(Long id) {
+		return this.netdiskFileDao.getMyselfAndChildFileId(id);
+	}
+
+	public Serializable[] getUserSharFileId2All(Long userId) {
+		return this.netdiskFileDao.getUserSharFileId2All(userId);
 	}
 }
